pyproject.toml
[project]
name = "rein"
version = "0.1.0"
description = "ReIN — harness your code. An open-source agentic coding runtime."
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "BDeMo" },
]
dependencies = [
    "anthropic>=0.39.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "websockets>=13.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
]

[project.scripts]
rein = "rein.main:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"
